firster rough draft for experienced user
The rev1.py script receives a "conditional" 2D sketch in DXF format (ASCII not Binary), which you previously export from your CAD application, and converts it to an encoding suitable for opening in CFTurbo.
The script works by clearing unnecessary data from the third coordinate system and sorting the remaining polyline vertices to minimize the number of "empty" transitions between them. When loading the converted sketch into CFTurbo, you'll receive a minimum number of "junk" polylines.

Debugger_renderer.py is used for debugging and is controlled through your IDE's console.
The sketches are provided as examples.